Changes
Changed the implementation of some classes to accomodate how sensors work.
Changed the MainActivity class to help and get a feel for how sensors work.
Added OnChange Thread Handler class, this will be how we will handle having multiple sensors registered to the SensorEventListener.
Finished and passed unittests for the vector classes.
Cleaned-up/Removed portions of the Service that are no longer needed.